fix: clean imports and initialize sync
