+++ server/app.js
@@ -1,151 +1,35 @@
 import express from "express";
-import fs from "fs";
 import path from "path";
 import { fileURLToPath } from "url";
-import Database from "better-sqlite3";
 import os from "os";
 import { format } from "date-fns";
+import db from "./lib/db.js";
+import {
+  initSync,
+  startSyncTimer,
+  setSyncRole,
+  setSyncServerUrl,
+  setSyncInterval,
+  setSyncEnabled,
+  setLlmUrl,
+  setLlmToken,
+  setLlmModel,
+  syncLogs,
+  getSyncRole,
+  getSyncStatus,
+  getLlmConfig,
+} from "./lib/sync.js";
+import { notifyClients, registerClient } from "./lib/sse.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
-const DATA_DIR = path.join(__dirname, "data");
-const DB_FILE = path.join(DATA_DIR, "data.db");
 const DIST_DIR = path.join(__dirname, "..", "dist");
-
-fs.mkdirSync(DATA_DIR, { recursive: true });
-const db = new Database(DB_FILE);
-
 function log(...args) {
   console.log(new Date().toISOString(), ...args);
 }
 
-db.exec(`
-  CREATE TABLE IF NOT EXISTS tasks (
-    id TEXT PRIMARY KEY,
-    data TEXT NOT NULL
-  );
-  CREATE TABLE IF NOT EXISTS categories (
-    id TEXT PRIMARY KEY,
-    data TEXT NOT NULL
-  );
-  CREATE TABLE IF NOT EXISTS notes (
-    id TEXT PRIMARY KEY,
-    data TEXT NOT NULL
-  );
-  CREATE TABLE IF NOT EXISTS recurring (
-    id TEXT PRIMARY KEY,
-    data TEXT NOT NULL
-  );
-  CREATE TABLE IF NOT EXISTS habits (
-    id TEXT PRIMARY KEY,
-    data TEXT NOT NULL
-  );
-  CREATE TABLE IF NOT EXISTS flashcards (
-    id TEXT PRIMARY KEY,
-    data TEXT NOT NULL
-  );
-  CREATE TABLE IF NOT EXISTS decks (
-    id TEXT PRIMARY KEY,
-    data TEXT NOT NULL
-  );
-  CREATE TABLE IF NOT EXISTS settings (
-    key TEXT PRIMARY KEY,
-    value TEXT NOT NULL
-  );
-  CREATE TABLE IF NOT EXISTS pomodoro_sessions (
-    start INTEGER NOT NULL,
-    end INTEGER NOT NULL,
-    breakEnd INTEGER
-  );
-  CREATE TABLE IF NOT EXISTS timers (
-    id TEXT PRIMARY KEY,
-    data TEXT NOT NULL
-  );
-  CREATE TABLE IF NOT EXISTS trips (
-    id TEXT PRIMARY KEY,
-    data TEXT NOT NULL
-  );
-  CREATE TABLE IF NOT EXISTS workdays (
-    id TEXT PRIMARY KEY,
-    data TEXT NOT NULL
-  );
-  CREATE TABLE IF NOT EXISTS inventory_items (
-    id TEXT PRIMARY KEY,
-    data TEXT NOT NULL
-  );
-  CREATE TABLE IF NOT EXISTS inventory_categories (
-    id TEXT PRIMARY KEY,
-    data TEXT NOT NULL
-  );
-  CREATE TABLE IF NOT EXISTS inventory_tags (
-    id TEXT PRIMARY KEY,
-    data TEXT NOT NULL
-  );
-  CREATE TABLE IF NOT EXISTS deletions (
-    type TEXT NOT NULL,
-    id TEXT NOT NULL,
-    deletedAt TEXT NOT NULL,
-    PRIMARY KEY (type, id)
-  );
-`);
-try {
-  db.prepare("ALTER TABLE pomodoro_sessions ADD COLUMN breakEnd INTEGER").run();
-} catch {}
 
-let syncRole = "client";
-let syncServerUrl = "";
-let syncInterval = 5;
-let syncEnabled = true;
-let llmUrl = "";
-let llmToken = "";
-let llmModel = "gpt-3.5-turbo";
-let syncTimer = null;
-let lastSyncTime = 0;
-let lastSyncError = null;
-const syncLogs = [];
-const sseClients = [];
-
-export function notifyClients() {
-  const msg = "data: update\n\n";
-  sseClients.forEach((res) => res.write(msg));
-}
-
-const initialSettings = loadSettings();
-if (typeof initialSettings.syncInterval === "number") {
-  syncInterval = initialSettings.syncInterval;
-}
-if (initialSettings.syncRole) {
-  syncRole = initialSettings.syncRole;
-}
-if (initialSettings.syncServerUrl) {
-  syncServerUrl = initialSettings.syncServerUrl;
-}
-if (typeof initialSettings.syncEnabled === "boolean") {
-  syncEnabled = initialSettings.syncEnabled;
-}
-if (initialSettings.llmUrl) {
-  llmUrl = initialSettings.llmUrl;
-}
-if (initialSettings.llmToken) {
-  llmToken = initialSettings.llmToken;
-}
-if (initialSettings.llmModel) {
-  llmModel = initialSettings.llmModel;
-}
-log("Initial sync settings", {
-  role: syncRole,
-  url: syncServerUrl,
-  interval: syncInterval,
-  enabled: syncEnabled,
-  llmConfigured: !!llmUrl,
-});
-startSyncTimer();
-try {
-  saveWorkDays(loadWorkDays());
-} catch (err) {
-  log("Failed to normalize workdays", err);
-}
 
 function dateReviver(key, value) {
   if (typeof value === "string" && /^\d{4}-\d{2}-\d{2}T/.test(value)) {
@@ -752,102 +636,41 @@ function applyDeletions(data) {
   return data;
 }
 
-async function performSync() {
-  if (syncRole !== "client" || !syncServerUrl) return;
-  const url = `${syncServerUrl.replace(/\/$/, "")}/api/sync`;
-  try {
-    log("Starting sync with", url);
-    const data = loadAllData();
-    if (data.settings) {
-      delete data.settings.syncServerUrl;
-      delete data.settings.syncRole;
-      delete data.settings.syncEnabled;
-      delete data.settings.llmToken;
-    }
-    const post = await fetch(url, {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify(data, (k, v) =>
-        v instanceof Date ? v.toISOString() : v,
-      ),
-    });
-    if (!post.ok) throw new Error(`HTTP ${post.status}`);
-    const res = await fetch(url);
-    if (!res.ok) throw new Error(`HTTP ${res.status}`);
-    const incoming = await res.json();
-    const merged = applyDeletions(mergeData(loadAllData(), incoming));
-    saveAllData(merged);
-    lastSyncTime = Date.now();
-    lastSyncError = null;
-    log("Sync successful");
-  } catch (err) {
-    console.error("Sync error", err);
-    lastSyncTime = Date.now();
-    lastSyncError = err.message || String(err);
-  }
+initSync({ loadAllData, saveAllData });
+const initialSettings = loadSettings();
+if (typeof initialSettings.syncInterval === "number") {
+  setSyncInterval(initialSettings.syncInterval);
 }
-
-function startSyncTimer() {
-  if (syncTimer) clearInterval(syncTimer);
-  syncTimer = null;
-  if (
-    syncEnabled &&
-    syncRole === "client" &&
-    syncServerUrl &&
-    syncInterval > 0
-  ) {
-    log("Sync timer started with interval", syncInterval, "minutes");
-    performSync();
-    syncTimer = setInterval(performSync, syncInterval * 60 * 1000);
-  }
+if (initialSettings.syncRole) {
+  setSyncRole(initialSettings.syncRole);
 }
-
-function setSyncRole(role) {
-  const newRole = role === "server" ? "server" : "client";
-  if (newRole === syncRole) return;
-  syncRole = newRole;
-  log("Sync role set to", syncRole);
-  startSyncTimer();
+if (initialSettings.syncServerUrl) {
+  setSyncServerUrl(initialSettings.syncServerUrl);
 }
-
-function setSyncServerUrl(url) {
-  if (url && !/^https?:\/\//i.test(url)) {
-    url = "http://" + url;
-  }
-  const normalized = url ? url.replace(/\/$/, "") : "";
-  if (normalized === syncServerUrl) return;
-  syncServerUrl = normalized;
-  log("Sync server URL set to", syncServerUrl || "(none)");
-  startSyncTimer();
+if (typeof initialSettings.syncEnabled === "boolean") {
+  setSyncEnabled(initialSettings.syncEnabled);
 }
-
-function setSyncInterval(minutes) {
-  const val = Number(minutes) || 0;
-  if (val === syncInterval) return;
-  syncInterval = val;
-  log("Sync interval set to", syncInterval);
-  startSyncTimer();
+if (initialSettings.llmUrl) {
+  setLlmUrl(initialSettings.llmUrl);
 }
-
-function setSyncEnabled(val) {
-  const enabled = Boolean(val);
-  if (enabled === syncEnabled) return;
-  syncEnabled = enabled;
-  log("Sync enabled set to", syncEnabled);
-  startSyncTimer();
+if (initialSettings.llmToken) {
+  setLlmToken(initialSettings.llmToken);
 }
-
-function setLlmUrl(url) {
-  llmUrl = url || "";
-  log("LLM URL set to", llmUrl || "(none)");
+if (initialSettings.llmModel) {
+  setLlmModel(initialSettings.llmModel);
 }
-
-function setLlmToken(token) {
-  llmToken = token || "";
-}
-
-function setLlmModel(model) {
-  llmModel = model || "gpt-3.5-turbo";
+log("Initial sync settings", {
+  role: getSyncRole(),
+  url: initialSettings.syncServerUrl || "",
+  interval: initialSettings.syncInterval || 0,
+  enabled: initialSettings.syncEnabled !== false,
+  llmConfigured: !!initialSettings.llmUrl,
+});
+startSyncTimer();
+try {
+  saveWorkDays(loadWorkDays());
+} catch (err) {
+  log("Failed to normalize workdays", err);
 }
 
 export const app = express();
@@ -860,12 +683,8 @@ app.get("/api/updates", (req, res) => {
     Connection: "keep-alive",
   });
   res.write("\n");
-  sseClients.push(res);
-  req.on("close", () => {
-    const idx = sseClients.indexOf(res);
-    if (idx !== -1) sseClients.splice(idx, 1);
+  registerClient(req, res);
   });
-});
 
 import createDataRouter from "./routes/data.js";
 import createFlashcardsRouter from "./routes/flashcards.js";
@@ -963,12 +782,12 @@ app.use(
     syncLogs,
     notifyClients,
     dateReviver,
-    syncRole: () => syncRole,
+    syncRole: getSyncRole,
   }),
 );
 app.use(
   "/api/sync-log",
-  createSyncLogRouter({ syncLogs, syncRole: () => syncRole }),
+  createSyncLogRouter({ syncLogs, syncRole: getSyncRole }),
 );
 let activePort = 3002;
 const publicIp = process.env.SERVER_PUBLIC_IP || null;
@@ -982,16 +801,12 @@ app.use(
 app.use(
   "/api/sync-status",
   createSyncStatusRouter({
-    getStatus: () => ({
-      last: lastSyncTime,
-      error: lastSyncError,
-      enabled: syncEnabled,
-    }),
+    getStatus: getSyncStatus,
   }),
 );
 app.use(
   "/api/llm",
-  createLlmRouter({ getConfig: () => ({ llmUrl, llmToken, llmModel }) }),
+  createLlmRouter({ getConfig: getLlmConfig }),
 );
 
 app.use(express.static(DIST_DIR));
